requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
Authlib==1.6.11
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudpickle==3.1.2
colorama==0.4.6
colorblind==0.0.9
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
docstring_parser==0.18.0
fastapi==0.136.0
filelock==3.28.0
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
google-adk==1.30.0
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-cloud-aiplatform==1.148.0
google-cloud-appengine-logging==1.9.0
google-cloud-audit-log==0.5.0
google-cloud-bigquery==3.41.0
google-cloud-bigquery-storage==2.37.0
google-cloud-bigtable==2.36.0
google-cloud-core==2.5.1
google-cloud-dataplex==2.18.0
google-cloud-discoveryengine==0.13.12
google-cloud-iam==2.22.0
google-cloud-logging==3.15.0
google-cloud-monitoring==2.30.0
google-cloud-pubsub==2.37.0
google-cloud-resource-manager==1.17.0
google-cloud-secret-manager==2.27.0
google-cloud-spanner==3.65.0
google-cloud-speech==2.38.0
google-cloud-storage==3.10.1
google-cloud-trace==1.19.0
google-crc32c==1.8.0
google-genai==1.73.1
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
graphviz==0.21
greenlet==3.4.0
grpc-google-iam-v1==0.14.4
grpc-interceptor==0.15.4
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.1
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
Mako==1.3.11
MarkupSafe==3.0.3
matplotlib==3.10.8
mcp==1.27.0
mmh3==5.2.1
moviepy==2.2.1
mpmath==1.3.0
multidict==6.7.1
narwhals==2.19.0
networkx==3.6.1
numpy==2.4.4
opencv-python==4.13.0.92
opentelemetry-api==1.38.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==26.1
pandas==3.0.2
pillow==11.3.0
polars==1.39.3
polars-runtime-32==1.39.3
proglog==0.1.12
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.1
pydantic-settings==2.13.1
pydantic_core==2.46.1
pydeck==0.9.2
PyJWT==2.12.1
pyOpenSSL==26.0.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
scipy==1.17.1
setuptools==81.0.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.49
sqlalchemy-spanner==1.17.3
sqlparse==0.5.5
sse-starlette==3.3.4
starlette==0.52.1
streamlit==1.56.0
sympy==1.14.0
tenacity==9.1.4
toml==0.10.2
torch==2.11.0
torchvision==0.26.0
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
ultralytics==8.4.38
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.44.0
watchdog==6.0.0
websockets==15.0.1
yarl==1.23.0
zipp==3.23.1
